PcapFile only template io method and add addPacket method
