--- libwire/inc/Hello.h
@@ -1,14 +1,0 @@
-#ifndef HELLO_H
-#define HELLO_H
-
-#include <string>
-
-class Hello {
-public:
-    Hello() = default;
-    ~Hello() = default;
-
-    std::string yo() const;
-};
-
-#endif // HELLO_H
